+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [unreleased]
+
+### Changed
+
+- React Native version to `0.72.6`
+- React Native peer dependency version to `0.65.0+`
+
 ## [0.14.2] (2023-11-27)
 
 ### Added
@@ -14,6 +21,7 @@
 - iOS: `onEvent` on iOS has incomplete payload information
 - tvOS: Picture in Picture sample screen has unwanted padding
 - iOS: hide home indicator when entering fullscreen mode in the example application
+- iOS: invalid `loadingState` value in `SeekEvent`, `SourceLoadEvent`, `SourceLoadedEvent` and in `SourceUnloadedEvent`
 
 ## [0.14.1] (2023-11-16)